pyproject.toml
[project]
name = "chunkcanvas"
version = "0.1.0"
description = "Document management system for AI assistants using Pinecone vector database"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2txt",
    "langchain-pinecone",
    "langchain-text-splitters",
    "langchain-voyageai",
    "llama-parse",
    "openpyxl",
    "pinecone-client",
    "PyPDF2",
    "streamlit"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/chunkcanvas"
"Bug Tracker" = "https://github.com/yourusername/chunkcanvas/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page1.py", "main.py"]

[tool.hatch.metadata]
allow-direct-references = true
